requirements.txt
# Core Application
fastapi>=0.115.0
uvicorn>=0.32.0
celery>=5.4.0
redis>=5.2.0

# Database & ORM
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.36
alembic>=1.13.3
pgvector>=0.3.7

# ML/CV Models
# Note: torch/torchvision versions may conflict with model packages (omnivinci requires 2.3.0)
# These can be installed separately if needed for specific models
torch>=2.3.0,<2.10.0
torchvision>=0.18.0,<0.25.0
transformers>=4.46.0
# Note: numpy 2.x may cause compatibility issues with some model packages
# OmniVinci requires numpy==1.26.4, Wildlife-tools works with 1.x
# Using <2.0 for broader compatibility
numpy>=1.26.0,<2.0.0
scipy>=1.14.0
scikit-learn>=1.6.0
Pillow>=11.0.0
opencv-python>=4.10.0
imageio>=2.35.0

# MCP & Agent Infrastructure
mcp>=0.9.0
markitdown>=0.0.1a2  # Document processing for evidence upload
firecrawl-py>=0.0.16

# Web Search (Free, no API key required)
ddgs>=9.0.0

# Modal for Serverless GPU Compute
modal>=1.2.0

# Data Processing
pandas>=2.2.3
openpyxl>=3.1.5
xlsxwriter>=3.2.0
pdfplumber>=0.11.0
python-docx>=1.1.2
markdown>=3.7.0
weasyprint>=62.0

# Authentication & Security
passlib>=1.7.4
bcrypt>=4.2.1
pyjwt>=2.9.0
python-jose>=3.3.0
python-multipart>=0.0.12
email-validator>=2.3.0

# Monitoring & Logging
prometheus-client>=0.21.0
opentelemetry-api>=1.27.0
structlog>=24.4.0
python-json-logger>=3.2.0
sentry-sdk>=2.20.0

# Anthropic Claude API
anthropic>=0.40.0

# Utilities
# Note: OmniVinci model package requires pydantic<2,>=1, but main app uses pydantic 2.x
# If using OmniVinci, install it separately with its own requirements
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
requests>=2.32.3
httpx>=0.27.2
openai>=1.0.0
google-genai>=1.0.0
aiofiles>=24.1.0

# Agent Frameworks
# Note: langchain 1.0+ is installed (current: 1.0.3), using >=0.3.0 for compatibility
langchain>=0.3.0
# Note: langgraph 1.0+ is installed (current: 1.0.2), using >=0.2.40 for compatibility  
langgraph>=0.2.40
autogen>=0.2.16

# Additional Utilities
python-dateutil>=2.9.0
pytz>=2024.2
pyyaml>=6.0.2

# Geocoding & Location Services (Phase 1: Location System Enhancement)
geopy>=2.3.0

# WebSocket Support (added for React frontend)
python-socketio>=5.11.0
aiohttp>=3.10.0
